refactor(anilist): split god class into layered files
Break the 1409 LOC AniListApi into a thin facade plus four
single-responsibility services under lib/core/api/anilist/
(transport, queries, media, MAL lookup, user lists). Drop
unused GraphQL fields (meanScore, popularity, season,
seasonYear, countryOfOrigin, nextAiringEpisode.airingAt) and
collapse the duplicated AniListAnimeGenreFilter into
AniListGenreFilter via a forAnime flag. Public API of
AniListApi is unchanged — no callers were touched.
