+++ lib/shared/models/manga.dart
@@ -99,8 +99,6 @@ class Manga {
       coverUrl: (coverMap?['extraLarge'] ?? coverMap?['large']) as String?,
       coverUrlMedium: coverMap?['medium'] as String?,
       averageScore: json['averageScore'] as int?,
-      meanScore: json['meanScore'] as int?,
-      popularity: json['popularity'] as int?,
       status: json['status'] as String?,
       startYear: dateMap?['year'] as int?,
       startMonth: dateMap?['month'] as int?,
@@ -108,7 +106,6 @@ class Manga {
       chapters: json['chapters'] as int?,
       volumes: json['volumes'] as int?,
       format: json['format'] as String?,
-      countryOfOrigin: json['countryOfOrigin'] as String?,
       genres: genresList?.map((dynamic g) => g as String).toList(),
       authors: authors,
       externalUrl: 'https://anilist.co/manga/$id',
